+++ controllers/tango_webapp/main_controller.js
@@ -19,7 +19,7 @@ TangoWebapp.MainController = class extends MVC.Controller{
 
         ui_builder.add_left_sidebar_item(TangoWebapp.ui.newDeviceInfoPanel(platform_api.context));
 
-        ui_builder.add_left_sidebar_item(TangoWebapp.ui.newDeviceControlPanel(platform_api.context));
+        ui_builder.add_left_sidebar_item(TangoWebapp.ui.newInfoControlPanel(platform_api.context));
 
         ui_builder.set_right_item(TangoWebapp.ui.newUserLogPanel(platform_api.context));
