+++ resources/webix_widgets/device_info_panel.js
@@ -1,0 +1,6 @@
+//TODO
+const device_info_panel = webix.protoUI({
+    name:"device_info_panel",
+    $init:function(){
+    }
+},  webix.ProgressBar, webix.ui.view);
