+++ resources/webix_widgets/import.js
@@ -5,6 +5,12 @@ import "./settings.js";
 import "./user_log.js";
 import "./info_datatable.js";
 import "./devices_tree.js";
+import "./tango_host_info_panel.js";
+import "./device_info_panel.js";
+import "./command_info_panel.js";
+import "./attr_info_panel.js";
+import "./pipe_info_panel.js";
+import "./info_control_panel.js";
 import "./device_view_panel.js";
 import "./plot.js";
 import "./scalar_view.js";
