+++ resources/webix_widgets/tango_host_info_panel.js
@@ -1,0 +1,6 @@
+//TODO
+const tango_host_info_panel = webix.protoUI({
+    name:"tango_host_info_panel",
+    $init:function(){
+    }
+},  webix.ProgressBar, webix.ui.view);
